chore: refactor, use and export calcPriceImpact function
