+++ packages/widget/src/index.ts
@@ -23,3 +23,4 @@ export { windows95Theme } from './themes/windows95.js';
 export * from './types/events.js';
 export type { TokenAmount } from './types/token.js';
 export * from './types/widget.js';
+export { calcPriceImpact } from './utils/calcPriceImpact.js';
